Rename profiles module to profiler
